movies/serializers.py
from rest_framework import serializers
from .models import Movie


class MovieFullSerializer(serializers.ModelSerializer):
    class Meta:
        model = Movie
        fields = ['id', 'title', 'description', 'year', 'poster']


class MovieSerializer(serializers.ModelSerializer):
    class Meta:
        model = Movie
        fields = ['id', 'poster']
