requirements.txt
Django==2.1.15
djangorestframework==3.11.0
Pillow==7.0.0
pytz==2019.3
django-cors-headers==3.2.1